server/requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
asyncpg 
blinker==1.9.0
boto3==1.40.60
botocore==1.40.60
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.79
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
Mako==1.3.10
MarkupSafe==3.0.3
narwhals==2.10.0
numpy==2.0.2
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pillow==11.3.0
pluggy==1.6.0
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.6.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.0.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
